refactor: split vault client types, add tests for reversible hashmap
